+++ hotel_reservation/utils.py
